Moved load dfd to superclass
